(feat)load images from the Mars Rover API
For each photo reference, we must load the actual image data using the photo reference's imageURL property.
In PhotosCollectionViewController.loadImage(forCell:, forItemAt:) implements
Get the MarsPhotoReference instance for the passed in indexPath from the photoReferences array property.
Get the URL for the associated image using the imageURL property. Use .usingHTTPS (provided in URL+Secure.swift) to make sure the URL is an https URL. By default, the API returns http URLs.
Create and run a data task to load the image data from the imageURL
In the data task's completion handler, check for an error, and return early if there is one. Otherwise, create a UIImage from the received data.
Check to see if the current index path for cell is the same one you were asked to load. If not, this means that that item has scrolled off screen and the UICollectionViewCell instance has been reused for a different index path. If this happens, abort setting the image.
If the cell hasn't been reused, set its imageView's image to the UIImage you just created.
Make sure all UIKit API calls on the main queue.
Initial commit.