(feat) Implements caching and thread-safety to address performance issues.
Instead of reloading the same images over and over again every time they're requested, we can store them in a cache after loading them the first time. Then, on subsequent requests for a given image, we check the cache, and if the image is already there we can avoid doing another network request.
a Cache class that is generic with respect to the type it stores, and its keys. e.g. Cache<Key, Value>, with Key type constrained to conform to Hashable.
a private property that is a dictionary to be used to store the cached UIImages. The generic type is [Key : Value]. Make sure you initialize it with an empty dictionary.
cache(value:, for:) and value(for:) methods to add items to the cache and remove them, respectively.
a cache property in PhotosCollectionViewController. Its keys are Ints, as we use MarsPhotoReference ids for the keys. Its values are UIImage objects, as we will be caching UIImages directly.
In  PhotosCollectionViewController.loadImage(forCell:, forItemAt:) method, before starting a data task, we first check to see if the cache already contains the UIImage for the given photo reference's id. If it exists, we set the cell's image immediately without doing a network request.
In the network request completion handler, we save the just-received image data to the cache so it is available later.
the Cache must be thread-safe by making sure that it ensures that all access to its underlying dictionary is done from only one thread at a time. We could use an instance of NSLock for this. However, since Grand Central Dispatch serial queues are designed to only allow one closure or unit of code to run at a time, they make for a great way to implement synchronized access to a shared resource without using a lock.
Create a private queue property and initialize it with a serial DispatchQueue.
In cache(value:, for:), dispatch the actual setting of the dictionary key/value pair so that it occurs on the queue.
In value(for:), use a synchronous dispatch to retrieve the requested value from the dictionary before returning it.
.