Implements image fetch cancellation via OperationQueue.
A major problem with the current implementation is that as you quickly scroll, every cell that scrolls on screen, no matter how briefly, triggers a network request. Even if a cell scrolls off-screen, its network request continues, preventing later cells' images from being loaded as quickly as possible. This also wastes CPU time, battery power, and network bandwidth loading images that the user may not actually look at.
To fix this, we will implement cancellation of in-flight network requests if they're no longer needed. This can be done directly with GCD, but Operation/OperationQueue are a natural fit for this kind of job.
Implementation:
-Implements PhotoFetchOperation, subclassed from ConcurrentOperation

-Updates PhotosCollectionViewController to use Operations

-Implements Cancellation in UICollectionViewDelegate method collectionView(_:, didEndDisplaying:, forItemAt:)
.